abstract network info
