+++ modules/network/kademlia/kademlia.js
@@ -13,6 +13,8 @@ const leveldown = require('leveldown');
 const PeerCache = require('./peer-cache');
 const ip = require('ip');
 
+const pjson = require('../../../package.json');
+
 const { NetworkRequestIgnoredError } = require('../../errors/index');
 
 /**
@@ -744,6 +746,28 @@ class Kademlia {
         }
 
         return true;
+    }
+
+    /**
+     * Returns basic network information
+     */
+    async getNetworkInfo() {
+        const peers = [];
+        const dump = this.node.router.getClosestContactsToKey(
+            this.node.identity,
+            kadence.constants.K * kadence.constants.B,
+        );
+
+        for (const peer of dump) {
+            peers.push(peer);
+        }
+
+        return {
+            versions: pjson.version,
+            identity: this.node.identity.toString('hex'),
+            contact: this.node.contact,
+            peers,
+        };
     }
 }
